requirements.txt
aiogram>=2.14.3
aiohttp>=3.7.4.post0
async-timeout>=3.0.1
attrs>=21.2.0
Babel>=2.9.1
certifi>=2021.5.30
chardet>=4.0.0
cycler>=0.10.0
greenlet>=1.1.0
idna>=3.2
kiwisolver>=1.3.1
matplotlib>=3.4.2
multidict>=5.1.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=8.3.1
psycopg2>=2.9.1
pyparsing>=2.4.7
python-dateutil>=2.8.2
pytz>=2021.1
six>=1.16.0
SQLAlchemy>=1.4.20
typing-extensions>=3.10.0.0
yarl>=1.6.3
